# golang-backend/docs/swagger.yaml
basePath: /
definitions:
  controller.BaseSwaggerResp:
    properties:
      code:
        example: 0
        type: integer
      msg:
        example: 操作成功
        type: string
      ts:
        example: 1700000000000
        type: integer
    type: object
...
  controller.SwaggerConfigGetReq:
    properties:
      name:
        example: app_name
        type: string
    type: object
  controller.SwaggerConfigGetResp:
    properties:
      code:
        example: 0
        type: integer
      data:
        example: some-value
        type: string
      msg:
        example: 操作成功
        type: string
      ts:
        example: 1700000000000
        type: integer
    type: object
...
  controller.SwaggerConfigUpdateMap:
    additionalProperties:
      type: string
    type: object
  controller.SwaggerConfigUpdateSingle:
    properties:
      name:
        example: app_name
        type: string
      value:
        example: network-panel
        type: string
    type: object
  controller.SwaggerForwardDiagnoseStepReq:
    properties:
      forwardId:
        example: 1
        type: integer
      step:
        example: entryExit
        type: string
    type: object
...
  controller.SwaggerIDReq:
    properties:
      id:
        example: 1
        type: integer
    type: object
...
  dto.ForwardDto:
    properties:
      inPort:
        type: integer
      interfaceName:
        type: string
      name:
        type: string
      remoteAddr:
        type: string
      strategy:
        type: string
      tunnelId:
        type: integer
    required:
    - name
    - remoteAddr
    - tunnelId
    type: object
  dto.ForwardUpdateDto:
    properties:
      id:
        type: integer
      inPort:
        type: integer
      interfaceName:
        type: string
      name:
        type: string
      remoteAddr:
        type: string
      strategy:
        type: string
      tunnelId:
        type: integer
    required:
    - id
    type: object
...
info:
  contact: {}
  description: 面板后端接口文档
  title: network-panel API
  version: "1.0"
paths:
  /api/v1/config/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: 配置名
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerConfigGetReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerConfigGetResp'
      summary: 根据 name 获取配置
      tags:
      - config
...
  /api/v1/config/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 键值对
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerConfigUpdateMap'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 批量更新配置
      tags:
      - config
  /api/v1/config/update-single:
    post:
      consumes:
      - application/json
      parameters:
      - description: name/value
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerConfigUpdateSingle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 单项配置更新
      tags:
      - config
  /api/v1/forward/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 转发信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ForwardDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 创建转发
      tags:
      - forward
  /api/v1/forward/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 转发ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 删除转发
      tags:
      - forward
  /api/v1/forward/diagnose:
    post:
      consumes:
      - application/json
      parameters:
      - description: 转发ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 诊断转发
      tags:
      - forward
  /api/v1/forward/diagnose-step:
    post:
      consumes:
      - application/json
      parameters:
      - description: 'forwardId, step: entryExit|nodeRemote|iperf3|path'
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerForwardDiagnoseStepReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 诊断转发分步
      tags:
      - forward
  /api/v1/forward/force-delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 转发ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 强制删除转发
      tags:
      - forward
...
  /api/v1/forward/pause:
    post:
      consumes:
      - application/json
      parameters:
      - description: 转发ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 暂停转发
      tags:
      - forward
  /api/v1/forward/resume:
    post:
      consumes:
      - application/json
      parameters:
      - description: 转发ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 恢复转发
      tags:
      - forward
  /api/v1/forward/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 转发更新参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ForwardUpdateDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 更新转发
      tags:
      - forward
  /api/v1/forward/update-order:
    post:
      consumes:
      - application/json
      parameters:
      - description: 转发顺序
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerForwardOrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 更新转发排序
      tags:
      - forward
  /api/v1/node/{id}/terminal:
    get:
      description: 仅管理员可用。建立 WS 后发送 JSON 消息：start/input/resize/stop。start 需携带 rows/cols，input
        携带 data。
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: integer
      - description: 备用token或Authorization
        in: query
        name: token
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
      summary: 节点终端 WebSocket
      tags:
      - ws
...
  /api/v1/node/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 创建节点
      tags:
      - node
  /api/v1/node/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点ID与是否卸载代理
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 删除节点
      tags:
      - node
  /api/v1/node/enable-gost-api:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeSimpleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 启用节点的gost API
      tags:
      - node
  /api/v1/node/get-exit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeSimpleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerResp'
      summary: 获取出口节点配置
      tags:
      - node
  /api/v1/node/gost-config:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeSimpleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerResp'
      summary: 获取节点上的gost配置
      tags:
      - node
  /api/v1/node/install:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeInstallReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerNodeInstallResp'
      summary: 获取节点安装命令
      tags:
      - node
  /api/v1/node/interfaces:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeInterfacesReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerResp'
      summary: 获取节点网卡IP列表
      tags:
      - node
  /api/v1/node/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 判定离线的阈值(毫秒)，默认30000
        in: query
        name: offline_threshold_ms
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerResp'
      summary: 节点列表
      tags:
      - node
  /api/v1/node/network-stats:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点ID与时间范围(1h/12h/1d/7d/30d)
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeRangeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerResp'
      summary: 节点连通性统计
      tags:
      - node
  /api/v1/node/network-stats-batch:
    post:
      consumes:
      - application/json
      parameters:
      - description: 时间范围(1h/12h/1d)
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNetworkStatsBatchReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerResp'
      summary: 批量节点网络统计
      tags:
      - node
  /api/v1/node/nq-result:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeSimpleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerResp'
      summary: 查询节点质量测试结果
      tags:
      - node
  /api/v1/node/nq-test:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeSimpleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerResp'
      summary: 触发节点质量测试
      tags:
      - node
  /api/v1/node/ops:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点或请求ID，可指定limit
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeOpsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerResp'
      summary: 查询节点操作日志
      tags:
      - node
  /api/v1/node/query-services:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点ID与过滤关键字
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeQueryServicesReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerResp'
      summary: 查询节点服务
      tags:
      - node
  /api/v1/node/restart-gost:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeSimpleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerResp'
      summary: 重启节点上的gost
      tags:
      - node
  /api/v1/node/set-exit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 出口服务配置
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeExitReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 配置出口节点服务
      tags:
      - node
  /api/v1/node/sysinfo:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点ID与时间范围
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeSysinfoReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerResp'
      summary: 获取节点系统信息
      tags:
      - node
  /api/v1/node/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 更新节点
      tags:
      - node
  /api/v1/nq/stream:
    post:
      consumes:
      - application/json
      parameters:
      - description: 回传内容
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeNQStreamReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: NodeQuality 流式回传
      tags:
      - node
  /api/v1/tunnel/bind/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerTunnelIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 获取隧道节点绑定IP
      tags:
      - tunnel
  /api/v1/tunnel/bind/set:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道ID与绑定IP
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerTunnelBindReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 设置隧道节点绑定IP
      tags:
      - tunnel
  /api/v1/tunnel/cleanup-temp:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerTunnelIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 清理路径临时服务（iperf3）
      tags:
      - tunnel
  /api/v1/tunnel/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.TunnelDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 创建隧道
      tags:
      - tunnel
  /api/v1/tunnel/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 删除隧道
      tags:
      - tunnel
  /api/v1/tunnel/diagnose:
    post:
      consumes:
      - application/json
      parameters:
      - description: tunnelId, step 可选 path|overlay|remote
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerTunnelDiagnoseReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 诊断隧道
      tags:
      - tunnel
  /api/v1/tunnel/diagnose-step:
    post:
      consumes:
      - application/json
      parameters:
      - description: 'tunnelId, step: entry|entryExit|exitPublic'
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerTunnelDiagnoseReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 隧道分步诊断
      tags:
      - tunnel
  /api/v1/tunnel/iface/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerTunnelIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 获取隧道节点出口IP
      tags:
      - tunnel
  /api/v1/tunnel/iface/set:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道ID与出口IP
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerTunnelIfaceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 设置隧道节点出口IP
      tags:
      - tunnel
...
  /api/v1/tunnel/path/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerTunnelIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 获取隧道多级路径
      tags:
      - tunnel
  /api/v1/tunnel/path/set:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道ID与路径节点
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerTunnelPathSetReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 设置隧道多级路径
      tags:
      - tunnel
  /api/v1/tunnel/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 隧道更新信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.TunnelUpdateDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 更新隧道
      tags:
      - tunnel
  /api/v1/tunnel/user/assign:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户隧道权限
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UserTunnelDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 分配隧道权限
      tags:
      - tunnel
  /api/v1/tunnel/user/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UserTunnelQueryDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 用户隧道权限列表
      tags:
      - tunnel
  /api/v1/tunnel/user/remove:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户隧道记录ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 移除用户隧道权限
      tags:
      - tunnel
...
swagger: "2.0"